refactor: merge unix directory permission files
- merge dir_permissions_core.go into dir_permissions_unix.go

- keep unix-only build constraint and syscall-based checks in one file

- remove duplicated cross-file split and simplify maintenance

- verify with go test -tags test for internal security packages

- verify with go build for record, verify, and runner binaries
